JARVIS-664 - Fix posting thumbnail runnable to UI thread
On some cases (older devices, scrolling fast) the runnable post to the
View (so it runs on UI thread) can actually fail to run.
The issue were more often observed on older devices and while scrolling
a list quickly.
The more reliable way of posting to UI thread now is to use the Handler.